// system-concierge/src/app/interfaces/user/iuser.ts
export interface IUser {

    companyId: number;
    resaleId: number;
    id: number;
    status: string;
    name: string;
    password: string;
    email: string;
    cellphone: string;
    limitDiscount: number;
    photo: string;
    roleId: number;
    roleDesc: string;
    roleFunc: string;

}